requirements.txt
Django==1.7
djangorestframework==3.3.1
wsgiref==0.1.2
